newport changes
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@34802 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/universal/linux-4.14/drivers/mtd/maps/nvram_kernel_x86.c
@@ -442,7 +442,7 @@ done:
 	return ret;
 }
 
-static int dev_nvram_ioctl(struct file *file, unsigned int cmd, unsigned long arg)
+static long dev_nvram_ioctl(struct file *file, unsigned int cmd, unsigned long arg)
 {
 	if (cmd != NVRAM_MAGIC) {
 //          printk(KERN_EMERG "Invalid nvram magic %X %X\n",cmd,NVRAM_MAGIC);
